Convert CrashDataStore to an interface for testing purpose.
